+++ CHANGELOG.md
@@ -7,7 +7,20 @@ e este projeto segue o [Semantic Versioning](https://semver.org/spec/v2.0.0.html
 
 ## [Unreleased]
 
-- Testes unitários e integração (não foram commitados, pois estou organizando os mesmos).
+- Testes de integração e carga.
+
+## [1.0.1]
+
+### [Added]
+
+- Testes unitários.
+- Criada exceção específica `UfNaoConfiguradaException` para indicar que uma uf não foi configurada.
+
+### [Changed]
+
+- As configurações foram implementadas como interface para permitir Mock.
+- O contrato de configuração foi alterado.
+- Alterada classe de parse para utilizar a nova exceção.
 
 ## [1.0.0]